refactor: split app domain state hooks
